Convert Acid to ARC. The NSMutableArray+InsertSort, JabberIQ, and JabberSession files are not for technical reasons.

+++ Frameworks/Acid/utility/esession/buffer.h
@@ -106,7 +106,7 @@ unsigned char* buffer_base64_encode(Buffer *buffer);
 		       ((u_int64_t)(u_char)(cp)[6] << 8) | \
 		       ((u_int64_t)(u_char)(cp)[7]))
 
-#define GET_32BIT(cp) (((u_long)(u_char)(cp)[0] << 24) | \
+#define GET_32BIT(cp) (u_int)(((u_long)(u_char)(cp)[0] << 24) | \
 		       ((u_long)(u_char)(cp)[1] << 16) | \
 		       ((u_long)(u_char)(cp)[2] << 8) | \
 		       ((u_long)(u_char)(cp)[3]))
